sub-wave-a-mcp-power-user-polish: DRY 33 handlers + registry state + 3 new tools
Three threads bundled (same files touched in multiple sub-tasks):

  A1. mcp_return_rpc_body_ctx() — extend the existing DRY helper with a
      printf-style parameter context so the rich error messages (name=%s,
      peer_id=%lld, txid=%s, etc.) survive the conversion. 33 manual
      7-line tails collapsed to 1-line returns across app/wallet/ops/
      chain/net controllers. Net delta: ~210 LOC removed. Side effects
      caught & fixed by reviewers:
        - chain_controller h_zcl_getblock: latent NULL-deref on bad height
        - wallet_controller h_zcl_listaddresses: same NULL-deref shape
        - wallet_controller h_zcl_replaywalletfromchain: missing
          early-return meant the destructive RPC ran even without
          confirm=true

  A2. zcl_state subsystem enum is now derived from the live g_dumpers
      registry in diagnostics_controller.c, instead of a hand-curated
      CSV that drifted (had 6 names; registry has 14). Added
      diagnostics_subsystems_csv() helper, populate p_state[0].enum_csv
      and description at mcp_register_ops() time. Adding a new
      *_dump_state_json subsystem now updates the MCP-visible schema
      with zero further plumbing.

  A3. Three new power-user MCP tools:
      - zcl_chain_tip: hash + height + time + age_seconds + work + bits
        + difficulty in one call (backed by new getchaintip RPC).
        Bundles getbestblockhash + getblockheader + chainwork.
      - zcl_mempool_inspect: fee-rate (zat/byte) and age histograms
        across mempool entries (backed by new getmempoolfeestats RPC).
        Power-user signal for fee construction and congestion diagnosis.
      - zcl_reorg_history: recent chain.reorg_* events (start,
        disconnect_failed, recovery_complete) via new getreorghistory
        RPC that wraps event_dump_json_filtered with the canonical
        prefix.
      EXPECTED_TOTAL bumped 88->91, EXPECTED_OPS 32->33,
      EXPECTED_CHAIN 11->13. (The 88/32 baseline was already drifted
      from the 87/31 the comment claimed.)

  + meta_controller.c p_admin.since bound tightened INT64_MAX -> INT32_MAX
    (unstaged before this session; folded in).

Three Explore agents mapped the scope; five general-purpose agents did
the parallel handler conversions per file with the new helper as a
single-source-of-truth reference.

Verification: full build, make lint clean, ./test_parallel 1/159 group
failures — the only remaining failure is the pre-existing
test_file_controller "consensus snapshot export" case (export guard
refuses to clobber downloaded snapshot when src utxos<1000, behaves
identically on baseline).

+++ tools/mcp/controllers/meta_controller.c
@@ -743,7 +743,7 @@ static const struct mcp_param_spec p_admin[] = {
      * embedded counters are cumulative since boot. */
     { "since", MCP_PARAM_INT, false,
       "Unix-seconds baseline for future windowed counters (unused).",
-      0, INT64_MAX, 0, 0, NULL, "0" },
+      0, INT32_MAX, 0, 0, NULL, "0" },
 };
 
 static const struct mcp_tool_route k_routes[] = {
